clean up conversation log
